feat: allow unknown feature flags if set to false
