+++ .eslint-dictionary.json
@@ -25,8 +25,8 @@
   "codebase",
   "codegen",
   "codepipeline",
-  "columnify",
   "cognito",
+  "columnify",
   "cors",
   "createCipheriv",
   "creds",
@@ -43,6 +43,7 @@
   "extensibility",
   "extentions",
   "facebook",
+  "filesystem",
   "foreach",
   "formatter",
   "frontend",
