+++ packages/amplify-cli-core/src/__tests__/testFiles/testProject-initialize-unknown-flag-false/cli.json
@@ -4,10 +4,7 @@
   },
   "features": {
     "graphQLTransformer": {
-      "customTransformerVersion": "it-is-a-string",
-      "foo": "unknown",
-      "bar": "unknown",
-      "disabled": false
+      "unknownFlagThatIsTurnedOff": false
     }
   }
 }
