+++ packages/amplify-cli-core/src/utils/index.ts
@@ -1,6 +1,9 @@
+export * from './doc-links';
 export * from './fileSize';
+/* eslint-disable import/no-cycle */
 export * from './isResourceNameUnique';
 export * from './open';
 export * from './packageManager';
 export * from './recursiveOmit';
+/* eslint-enable */
 export * from './validate-path';
