+++ packages/amplify-cli-npm/run.ts
@@ -1,5 +1,4 @@
 #!/usr/bin/env node
-
 import { run } from '.';
 
 run();
